Renamed file and removed few commented lines
